ManualFlow_Core/requirements.txt
djangorestframework==3.16.0
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-admin-interface==0.30.0
Django==5.1.4
python-dotenv==1.0.1